+++ MARS_NB/MARS/MARSCore/src/mars/sensors/LaserScanner.java
@@ -95,7 +95,7 @@ public class LaserScanner extends RayBasedSensor {
      */
     @Override
     protected float calculateAverageNoiseFunction(float x) {
-        return ((float) Math.pow(1.1f, (float) Math.abs(x)));
+        return ((float) Math.pow(1.1f, Math.abs(x)));
     }
 
     /**
@@ -105,6 +105,6 @@ public class LaserScanner extends RayBasedSensor {
      */
     @Override
     protected float calculateStandardDeviationNoiseFunction(float x) {
-        return ((float) Math.pow(1.1f, (float) Math.abs(x)));
+        return ((float) Math.pow(1.1f, Math.abs(x)));
     }
 }
